Reworked whole assignment - students need to interact more with interfaces

+++ src/main/java/h05/AbstractDurableObject.java
@@ -2,36 +2,24 @@ package h05;
 
 import org.tudalgo.algoutils.student.annotation.SolutionOnly;
 
-/**
- * A skeletal implementation of the {@link Durable} interface, intended to simplify
- * the creation of durable objects within the game.
- *
- * <p>Subclasses can focus on game-specific behavior without needing to reimplement
- * durability management.
- *
- * @author Nhan Huynh, Nico Schnieders
- */
 @SolutionOnly
 public class AbstractDurableObject implements Durable {
 
-    /**
-     * The current durability of the object in percent. Guaranteed to be in the range [0, 100].
-     */
-    private int durability = 100;
+    private double durability = 100;
 
     @Override
-    public int getDurability() {
+    public double getDurability() {
         return durability;
     }
 
     @Override
-    public void setDurability(int durability) {
+    public void setDurability(double durability) {
         this.durability = durability > 100 ? 100 : durability < 0 ? 0 : durability;
     }
 
     @Override
-    public void reduceDurability(int amount) {
-        int newDurability = durability - amount;
+    public void reduceDurability(double amount) {
+        double newDurability = durability - amount;
         durability = newDurability < 0 ? 0 : newDurability;
     }
 }
